Fail tests if list is a single node or a 4 element node with cycle at node 1
