+++ src/main/java/com/lperthel/datastructures/Solution.java
@@ -1,18 +1,8 @@
 package com.lperthel.datastructures;
 
 public class Solution {
-
-public boolean hasCycle(ListNode head) {
-	if(head ==null)
-		return false;
-	ListNode scout = head.next;
-	while(head!=scout && head!=null && scout!=null && scout.next!=null) {
-		head= head.next;
-		scout=scout.next.next;
-	}
-	if(head ==scout)
-		return true;
-	else
-		return false;
+    public ListNode detectCycle(ListNode head) {
+        return head;
     }
+
 }
